# openapi.yaml
openapi: 3.0.1
info:
  title: gpt-db API
  version: 1.0.0
servers:
  - url: https://gpt-db.vercel.app
    description: Production server
paths:
...
  /food/log:
    get:
      summary: Retrieve log entries
      operationId: getFoodLog
      tags:
        - food
      parameters:
        - in: query
          name: date
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
    post:
      summary: Append log entry
      operationId: appendFoodLog
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Log entry created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  log_id:
                    type: string
  /food/log/{log_id}:
    delete:
      summary: Delete log entry
      operationId: deleteFoodLog
      tags:
        - food
      parameters:
        - in: path
          name: log_id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Log entry deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  deleted:
                    type: boolean
...
  /food/targets:
    get:
      summary: Get macro targets
      operationId: getFoodTargets
      tags:
        - food
      responses:
        '200':
          description: Current targets
    patch:
      summary: Update macro targets
      operationId: patchFoodTargets
      tags:
        - food
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                calories:
                  type: number
                  minimum: 0
                  description: Target calories
                protein:
                  type: number
                  minimum: 0
                  description: Target protein grams
                fat:
                  type: number
                  minimum: 0
                  description: Target fat grams
                carbs:
                  type: number
                  minimum: 0
                  description: Target carbohydrate grams
              additionalProperties: false
      responses:
        '200':
          description: Targets updated
    delete:
      summary: Reset all macro targets
      operationId: deleteFoodTargets
      tags:
        - food
      responses:
        '200':
          description: Targets reset
  /food/targets/{macro}:
    delete:
      summary: Reset one macro target
      operationId: deleteFoodTargetsMacro
      tags:
        - food
      parameters:
        - in: path
          name: macro
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Target reset
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
security:
  - ApiKeyAuth: []